This adds auto-adjustable brightness.
State (on/off) is stored in a global variable. There's a switch to
enable/disable the function from HA.

Brightness is calculated from the ambient light sensor. Looking at the
illuminance-table at
wikipedia, it seems reasonable to divide ambient light by 4 and add a
bit, so the algorithm I've used is: `x / 4 + aab_add`, where `aab_add`
is a variable you set. Also, to avoid setting a new brightness every 10
seconds, we only adjust brightness if it is off by more than 2%.